requirements.txt
Python==3.9.13
asgiref==3.2.10
certifi==2022.6.15
charset-normalizer==2.1.0
crispy-bootstrap5==0.6
Django==3.1.2
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
wincertstore==0.2
